+++ src/app/file-chooser/file-chooser.component.html
@@ -3,4 +3,5 @@
 <!-- Cannot use the "accept" attribute as there is not a mime type for YAML -->
 <input type="file"
        id="api-definition-fc" name="api-definition-fc"
+       multiple="multiple"
        (change)="loadFile($event)">
